requirements.txt
telethon==1.30.3
python-dotenv==1.0.0
